+++ R/get_uuid.R
@@ -7,6 +7,9 @@
 #' @param name \code{character}. A taxonomic name. Various taxonomic levels
 #'   are supported (e.g. species, genus, family). NULL can also be supplied
 #'   which will skip the taxonomic filtering of the PhyloPic database.
+#' @param img A [Picture][grImport2::Picture-class] or png array object from 
+#'   [get_phylopic()]. A list of these objects can also be supplied. If `img`
+#'   is supplied, `name` and `n` are ignored. Defaults to NULL.
 #' @param n \code{numeric}. How many uuids should be returned? Depending on
 #'   the requested `name`, multiple silhouettes might exist. If `n` exceeds
 #'   the number of available images, all available uuids will be returned.
@@ -26,7 +29,26 @@
 #' @examples
 #' uuid <- get_uuid(name = "Acropora cervicornis")
 #' uuid <- get_uuid(name = "Dinosauria", n = 5, url = TRUE)
-get_uuid <- function(name = NULL, n = 1, url = FALSE) {
+get_uuid <- function(name = NULL, img = NULL, n = 1, url = FALSE) {
+  # Handle img -----------------------------------------------------------
+  if (!is.null(img)) {
+    if (is.list(img)) {
+      uuid <- sapply(img, function(x) attr(x, "uuid"))
+    } else {
+      uuid <- attr(img, "uuid")
+    }
+    if (any(is.null(uuid))) {
+      stop("uuid not available. Check `img` is from get_phylopic.")
+    }
+    if (url) {
+      if (is.list(img)) {
+        uuid <- sapply(img, function(x) attr(x, "url"))
+      } else {
+        uuid <- attr(img, "url")
+      }
+    }
+    return(uuid)
+  }
   # Error handling -------------------------------------------------------
   if (!is.null(name) && !is.character(name)) {
     stop("`name` should be `NULL` or of class character.")