+++ tests/testthat/test-get_attribution.R
@@ -8,9 +8,14 @@ test_that("get_attribution works", {
   # Expect equal
   uuid <- get_uuid(name = "Scleractinia", n = 5)
   expect_equal(length(get_attribution(uuid = uuid)), 5)
+  expect_message(get_attribution(uuid = uuid, text = TRUE))
+  # Check img arg
+  img <- get_phylopic(uuid = uuid[1])
+  expect_true(is.list(get_attribution(img = img)))
 
   # Expect error
   expect_error(get_attribution(uuid = NULL))
   expect_error(get_attribution(uuid = 1))
   expect_error(get_attribution(uuid = uuid, text = 1))
+  expect_error(get_attribution(img = 1))
 })